requirements.txt
# Django & API
Django>=5.1,<6.0
djangorestframework>=3.15
django-filter>=24.0
django-cors-headers>=4.3
django-environ>=0.11
drf-spectacular>=0.27  # OpenAPI/Swagger docs

# Task Queue
celery[redis]>=5.4
django-celery-beat>=2.6
django-celery-results>=2.5

# Database
psycopg2-binary>=2.9
django-pgconnection>=1.0

# HTTP & Scraping
httpx>=0.27
beautifulsoup4>=4.12
lxml>=5.1
selectolax>=0.3

# GitHub
PyGithub>=2.3
gitpython>=3.1

# AI / LLM
anthropic>=0.40
openai>=1.50

# Utilities
python-dateutil>=2.9
rich>=13.7  # Pretty console output
click>=8.1  # CLI commands
pydantic>=2.6  # Data validation
structlog>=24.1  # Structured logging
croniter>=2.0

# Testing
pytest>=8.0
pytest-django>=4.8
pytest-asyncio>=0.23
factory-boy>=3.3
httpx  # for async test client
coverage>=7.4
